feat: add Dashboard page and integrate it into the routing; update navigation links
